+++ src/main/kotlin/it/czerwinski/intellij/wavefront/lang/completion/MtlOptionCompletionProvider.kt
@@ -1,0 +1,89 @@
+/*
+ * Copyright 2020-2023 Slawomir Czerwinski
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package it.czerwinski.intellij.wavefront.lang.completion
+
+import com.intellij.codeInsight.completion.CompletionParameters
+import com.intellij.codeInsight.completion.CompletionProvider
+import com.intellij.codeInsight.completion.CompletionResultSet
+import com.intellij.codeInsight.lookup.LookupElementBuilder
+import com.intellij.util.ProcessingContext
+import it.czerwinski.intellij.wavefront.WavefrontObjBundle
+
+internal class MtlOptionCompletionProvider : CompletionProvider<CompletionParameters>() {
+
+    override fun addCompletions(
+        parameters: CompletionParameters,
+        context: ProcessingContext,
+        result: CompletionResultSet
+    ) {
+        result.addAllElements(
+            options.map { element ->
+                LookupElementBuilder.create(element.text)
+                    .withTypeText(element.description)
+            }
+        )
+    }
+
+    companion object {
+        private val options = listOf(
+            CompletionElement(
+                text = "blendu",
+                description = WavefrontObjBundle.message(key = "fileTypes.mtl.tokenType.BLEND_U_OPTION_NAME")
+            ),
+            CompletionElement(
+                text = "blendv",
+                description = WavefrontObjBundle.message(key = "fileTypes.mtl.tokenType.BLEND_V_OPTION_NAME")
+            ),
+            CompletionElement(
+                text = "cc",
+                description = WavefrontObjBundle.message(key = "fileTypes.mtl.tokenType.COLOR_CORRECTION_OPTION_NAME")
+            ),
+            CompletionElement(
+                text = "clamp",
+                description = WavefrontObjBundle.message(key = "fileTypes.mtl.tokenType.CLAMP_OPTION_NAME")
+            ),
+            CompletionElement(
+                text = "imfchan",
+                description = WavefrontObjBundle.message(key = "fileTypes.mtl.tokenType.SCALAR_CHANNEL_OPTION_NAME")
+            ),
+            CompletionElement(
+                text = "mm",
+                description = WavefrontObjBundle.message(key = "fileTypes.mtl.tokenType.VALUE_MODIFIER_OPTION_NAME")
+            ),
+            CompletionElement(
+                text = "o",
+                description = WavefrontObjBundle.message(key = "fileTypes.mtl.tokenType.OFFSET_OPTION_NAME")
+            ),
+            CompletionElement(
+                text = "s",
+                description = WavefrontObjBundle.message(key = "fileTypes.mtl.tokenType.SCALE_OPTION_NAME")
+            ),
+            CompletionElement(
+                text = "t",
+                description = WavefrontObjBundle.message(key = "fileTypes.mtl.tokenType.TURBULENCE_OPTION_NAME")
+            ),
+            CompletionElement(
+                text = "texres",
+                description = WavefrontObjBundle.message(key = "fileTypes.mtl.tokenType.RESOLUTION_OPTION_NAME")
+            ),
+            CompletionElement(
+                text = "bm",
+                description = WavefrontObjBundle.message(key = "fileTypes.mtl.tokenType.BUMP_MULTIPLIER_OPTION_NAME")
+            )
+        )
+    }
+}